+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # QOwnNotes Changelog
 
+## 22.6.0
+- users are now asked if they want to turn of the search panel again when using
+  the *Find note* action if they have turned it off previously
+  (for [#2548](https://github.com/pbek/QOwnNotes/issues/2548))
+
 ## 22.5.3
 - the note preview will now not being updated while typing by default
   (for [#2535](https://github.com/pbek/QOwnNotes/issues/2535))